Полный рафакторинг всего проекта по причине ошибок во время выполнения запросов на поиск событий с фильтрами у пользователей и админов

+++ ewm/src/main/java/ru/practicum/ewm/user/service/UserServiceImpl.java
@@ -1,21 +1,21 @@
 package ru.practicum.ewm.user.service;
 
 import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.ewm.error.exception.user.UserExistException;
-import ru.practicum.ewm.user.dto.UserRequestDto;
+import ru.practicum.ewm.error.exception.EntityConflictException;
+import ru.practicum.ewm.error.exception.EntityNotFoundException;
+import ru.practicum.ewm.user.dto.NewUserRequest;
+import ru.practicum.ewm.user.dto.UserDto;
 import ru.practicum.ewm.user.mapper.UserMapper;
 import ru.practicum.ewm.user.repository.UserRepository;
 
 import java.util.List;
 
-@Slf4j
 @Service
+@Transactional
 @AllArgsConstructor
-@Transactional(readOnly = true)
 public class UserServiceImpl implements UserService {
 
     private final UserRepository repository;
@@ -23,38 +23,40 @@ public class UserServiceImpl implements UserService {
     private final UserMapper mapper;
 
     @Override
+    @Transactional(readOnly = true)
+    public List<UserDto> getAllUsers(List<Long> ids, int from, int size) {
+
+        var users = ids == null || ids.isEmpty() ?
+                repository.findAll(PageRequest.of(from / size, size)).toList()
+                : repository.findAllById(ids);
+
+        return mapper.toDtoList(users);
+    }
+
+    @Override
     @Transactional
-    public UserRequestDto save(UserRequestDto user) {
+    public UserDto save(NewUserRequest user) {
 
         if (repository.existsUserByName(user.getName())) {
 
-            log.error("User " + user.getName() + "already exist.");
-
-            throw new UserExistException("User " + user.getName() + "already exist.");
+            throw new EntityConflictException("Пользователь с именем " + user.getName() + " уже существует.");
         }
 
-        return mapper.toUserDto(
-                repository.save(
-                        mapper.toUser(user)));
+        var response = repository.save(mapper.toUser(user));
+
+        return mapper.toUserDto(response);
     }
 
     @Override
-    public List<UserRequestDto> get(List<Long> ids, int from, int size) {
-
-        if (ids != null && ids.size() > 0) {
-
-            return mapper.toDtoList(repository.findAllById(ids));
-        } else {
-
-            return mapper.toDtoList(repository.findAll(PageRequest.of(from / size, size)).toList());
-
-        }
-    }
-
-    @Override
-    @Transactional
     public void delete(Long id) {
 
-        repository.deleteById(id);
+        if (repository.existsById(id)) {
+
+            repository.deleteById(id);
+
+            return;
+        }
+
+        throw new EntityNotFoundException("Ошибка удаления. Пользователь с id " + id  + " не найден.");
     }
 }
